+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Added `decode_hash`, `regexp` and `original` decoders
 * Fixed incorrect `Decoding::Failure` value comparisons
+* Fixed some decoders incorrectly returning `String` values rather than
+  `Decoding::Failure` values
 
 ## [0.1.0] - 2024-05-08